+++ CHANGELOG.md
@@ -22,5 +22,6 @@
 - Fix prompt being overwritten when entering a long line of text to the shell.
 - add (admitedly horendous) SIGINT (CTRL+C) handling so the shell kills the child process but keeps itself running.
 - Release v0.2-alpha!
+- add `history` built-in command.
 
 **